requirements.txt
aiohttp>=3.8.3
numpy>=1.23.5
python-dotenv>=0.21.0
pydantic>=1.10.2
websockets>=10.4
asyncio>=3.4.3
pandas>=1.5.2
